Split grpc server and client into separate files
